% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auth_gs}
\alias{auth_gs}
\title{Authenticate with Google Sheets}
\usage{
auth_gs(email = NULL)
}
\arguments{
\item{email}{The email to authenticate with Google Sheets. If not provided,
manual authentication will be requested.}
}
\description{
Authentication is required to access Google Sheets via the 'googlesheets4' package.
}
