% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_datasets.R
\name{dl_datasets}
\alias{dl_datasets}
\title{Download datasets required to update raw datasets for CovidTimelineCanada}
\usage{
dl_datasets()
}
\value{
A character vector containing the path of the temporary directory
containing the datasets required for \code{\link{update_raw_datasets}}. Names correspond to
the UUIDs assigned to the dataset in datasets.json.
}
\description{
The live version of each required dataset is downloaded from a URL using the
function \code{\link[Covid19CanadaData]{dl_dataset}}. Datasets are saved as
.RData files and HTML files to a temporary directory. Datasets are referred
to by their UUID given in the datasets.json (\url{https://github.com/ccodwg/Covid19CanadaArchive/blob/master/datasets.json})
file from "Covid19CanadaArchive" (\url{https://github.com/ccodwg/Covid19CanadaArchive}).
}
