% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_covid19tracker_d}
\alias{get_covid19tracker_d}
\title{Get covid19tracker.ca data for a particular value and region}
\usage{
get_covid19tracker_d(val, region, from = NULL, to = NULL)
}
\arguments{
\item{val}{The value to read data for. One of "hospitalizations" or "icu".}

\item{region}{Either "all" or the region to read data for.}

\item{from}{Optional. Exclude data prior to this date.}

\item{to}{Optional. Exclude data after this date.}
}
\description{
Get covid19tracker.ca data for a particular value and region
}
