% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_hr_data}
\alias{get_hr_data}
\title{Get health region data}
\usage{
get_hr_data(sknew = TRUE)
}
\arguments{
\item{sknew}{Use the new SK health regions? Default: True.}
}
\value{
A data frame containing the information from hr_map.csv.
}
\description{
Get health region data
}
