% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_phac_d}
\alias{get_phac_d}
\title{Get PHAC data for a particular value and region}
\usage{
get_phac_d(val, region, exclude_repatriated = TRUE)
}
\arguments{
\item{val}{The value to read data for.}

\item{region}{Either "all" or the region to read data for.}

\item{exclude_repatriated}{Exclude "Repatriated" region when reading data for "all" regions? Default: TRUE.}
}
\description{
Get PHAC data for a particular value and region
}
