% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_datasets.R
\name{load_ds}
\alias{load_ds}
\title{Load data from temporary directory}
\usage{
load_ds(ds, ds_name, d)
}
\arguments{
\item{ds}{The list of datasets returned by \code{\link[Covid19CanadaETL]{dl_datasets}}.}

\item{ds_name}{The UUID of the dataset to load.}

\item{d}{The object already loaded into memory. If provided, all other arguments will be ignored.}

\item{skip_raw_update}{If TRUE, skip updating raw datasets and just assemble final datasets. Default: FALSE.}
}
\description{
Load data from temporary directory
}
