% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_datasets.R
\name{plot_datasets}
\alias{plot_datasets}
\title{Plot CovidTimelineCanada datasets}
\usage{
plot_datasets(plot_dir = ".")
}
\arguments{
\item{plot_dir}{Directory to save plots. Defaults to current directory.}
}
\description{
This function expects that data have already been loaded using \code{\link[Covid19CanadaETL]{load_datasets}}.
}
