% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{pushover}
\alias{pushover}
\title{Send Pushover notification via POST request to GitHub Actions API}
\usage{
pushover(message, priority, title, device)
}
\arguments{
\item{message}{The body of the notification.}

\item{priority}{Optional. Message priority, an integer from -2 to 2, see: https://pushover.net/api#priority. Defaults to 0 (normal priority).}

\item{title}{Optional. The title of the notification. If missing, the application's name will be used.}

\item{device}{Optional. The name of the device to send the notification to. If missing, the notification will be sent to all devices.}
}
\description{
Environmental variable `GITHUB_PAT` must be present.
}
