% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_hr}
\alias{replace_hr}
\title{Replace part of a health region-level time series}
\usage{
replace_hr(d1, d2, val, region, date)
}
\arguments{
\item{d1}{The health region-level time series to replace data in.}

\item{d2}{The replacement data for a specific region.}

\item{val}{The value to replace. One of "cases" or "deaths".}

\item{region}{The region to replace data for.}

\item{date}{The date to start replacing data from.}
}
\description{
Replace part of a health region-level time series
}
