% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_hr_phac}
\alias{replace_hr_phac}
\title{Replace part of health region-level time series with PHAC data}
\usage{
replace_hr_phac(d, val, region, date)
}
\arguments{
\item{d}{The health region-level time series to replace data in.}

\item{val}{The value to replace. One of "cases" or "deaths".}

\item{region}{The region to replace data for.}

\item{date}{The date to start replacing data from.}
}
\description{
Replace part of health region-level time series with PHAC data
}
