% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messaging.R
\name{send_email}
\alias{send_email}
\title{Send email via POST request to GitHub Actions API}
\usage{
send_email(subject, body)
}
\arguments{
\item{subject}{Email subject.}

\item{body}{Email body.}
}
\description{
Environmental variable `GITHUB_PAT` must be present.
}
