% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_funs.R
\name{sheets_funs}
\alias{sheets_funs}
\alias{auth_gd}
\alias{auth_gs}
\alias{sheets_load}
\alias{sheets_merge}
\alias{sheets_upload}
\alias{upload_active_cumul}
\title{Functions to download/upload data from Google Sheets}
\usage{
auth_gd(email = NULL, path = NULL)

auth_gs(email = NULL, path = NULL)

sheets_load(files, file, sheet = NULL)

sheets_merge(d, d_merge)

sheets_upload(d, files, file, sheet = NULL, col_range = NULL)

upload_active_cumul(d, files, file, sheet = NULL)
}
\arguments{
\item{email}{The email to authenticate with Google Sheets/Drive. If not
provided, manual authentication will be requested (unless `path` is provided).}

\item{path}{Alternative to email authentication. JSON identifying service
account (e.g., file path to JSON). See \code{\link[googledrive]{drive_auth}}
and \code{\link[googlesheets4]{gs4_auth}}.}

\item{files}{List of Google Drive files from \code{\link[googledrive]{drive_ls}}.}

\item{file}{Name of specific Google Sheets file.}

\item{sheet}{Number/name of specific sheet in the Google Sheets file.}

\item{d}{Dataset to upload.}

\item{d_merge}{Dataset to merge into.}

\item{col_range}{Column range to write to (e.g., "C:C").}
}
\description{
Functions to download/upload data from Google Sheets

Authenticate with Google Drive

Authenticate with Google Sheets

Quickly load data from Google Sheets

Merge data with existing data in Google Sheets

Upload data to Google Sheets

Upload active_cumul data to Google Sheets
}
