% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_datasets.R
\name{update_active_cumul}
\alias{update_active_cumul}
\title{Update active_cumul datasets}
\usage{
update_active_cumul(ds)
}
\arguments{
\item{ds}{The list of datasets returned by \code{\link[Covid19CanadaETL]{dl_datasets}}.}
}
\description{
Update active_cumul datasets
}
