% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_datasets.R
\name{update_active_ts}
\alias{update_active_ts}
\title{Update active_ts datasets}
\usage{
update_active_ts(ds)
}
\arguments{
\item{ds}{The list of datasets returned by \code{\link[Covid19CanadaETL]{dl_datasets}}.}
}
\description{
Update active_ts datasets
}
