% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_raw_datasets.R
\name{update_raw_datasets}
\alias{update_raw_datasets}
\title{Update raw datasets for CovidTimelineCanada}
\usage{
update_raw_datasets(skip_raw_update = FALSE)
}
\description{
Update raw datasets for CovidTimelineCanada
}
