% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_time.R
\name{update_time}
\alias{update_time}
\title{Write update time}
\usage{
update_time()
}
\value{
The update time as a string.
}
\description{
The format of the update time is %Y-%m-%d %H:%M %Z and the time zone is
America/Toronto.
}
