% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_funs.R
\name{write_funs}
\alias{write_funs}
\alias{write_ts}
\alias{update_covid19tracker_dataset}
\alias{sync_report}
\alias{sync_active_cumul}
\alias{write_dataset}
\title{Functions to write data}
\usage{
write_ts(d, dir = c("active_ts", "active_cumul"), region, val)

update_covid19tracker_dataset(val)

sync_report(sheet, region, geo)

sync_active_cumul(sheet, val, regions)

write_dataset(d, geo, name)
}
\arguments{
\item{d}{The dataset to write.}

\item{dir}{The directory to write to. One of "active_ts", "active_cumul".}

\item{region}{The region to write data for.}

\item{val}{The name identifying the value being written.}

\item{sheet}{Name of sheet to read from the reports or active_cumul Google Sheet.}

\item{geo}{The geographic level of data being written. One of "can", "pt", "hr", "sub-hr".}

\item{regions}{Regions to sync from active_cumul Google Sheet.}

\item{name}{The name to give to the output file.}
}
\description{
Functions to write data

Write processed data to a specific directory in "raw_data"

Download and write covid19tracker.ca hospitalization or ICU data to "raw_data/covid19tracker" directory

Sync and write report data from Google Sheets

Sync and write active_cumul data from Google Sheets

Write final CSV dataset
}
