% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccodwg_update.R
\name{ccodwg_update}
\alias{ccodwg_update}
\title{Perform update of CovidTimelineCanada}
\usage{
ccodwg_update(email = NULL, path = NULL)
}
\arguments{
\item{email}{The email to authenticate with Google Sheets. If not provided,
manual authentication will be requested.}

\item{path}{Alternative to email authentication. JSON identifying service
account (e.g., file path to JSON). See \code{\link[googlesheets4]{gs4_auth}}.}
}
\description{
Perform update of CovidTimelineCanada
}
