% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccodwg_update.R
\name{ccodwg_update}
\alias{ccodwg_update}
\title{Perform update of CovidTimelineCanada}
\usage{
ccodwg_update(
  email = NULL,
  path = NULL,
  skip_raw_update = FALSE,
  skip_extra_update = FALSE
)
}
\arguments{
\item{email}{The email to authenticate with Google Sheets. If not provided,
manual authentication will be requested.}

\item{path}{Alternative to email authentication. JSON identifying service
account (e.g., file path to JSON). See \code{\link[googlesheets4]{gs4_auth}}.}

\item{skip_raw_update}{If TRUE, skip updating raw datasets and just assemble final datasets. Default: FALSE.}

\item{skip_extra_update}{If TRUE, skip updating extra datasets. Default: FALSE.}
}
\description{
Perform update of CovidTimelineCanada
}
