% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_t_datasets.R
\name{e_t_datasets}
\alias{e_t_datasets}
\title{Extract and transform data for the daily Canadian COVID-19 data collection}
\usage{
e_t_datasets(mode = c("main", "phu"))
}
\arguments{
\item{mode}{Process which dataset? One of "main" or "phu".}
}
\description{
Extract and transform data for the daily Canadian COVID-19 data collection
}
