% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_datasets.R
\name{extra_datasets}
\alias{extra_datasets}
\title{Assemble and write extra datasets for CovidTimelineCanada}
\usage{
extra_datasets(skip_extra_update = FALSE)
}
\arguments{
\item{skip_extra_update}{If TRUE, skip updating extra datasets. Default: FALSE.}
}
\description{
Assemble and write extra datasets for CovidTimelineCanada
}
