% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{process_funs}
\alias{process_funs}
\alias{convert_hr_names}
\alias{date_shift}
\alias{add_name_col}
\alias{drop_sub_regions}
\alias{add_as_of_date}
\alias{report_pluck}
\alias{report_recent}
\alias{max_date}
\alias{append_daily_d}
\alias{collate_datasets}
\alias{add_hr_col}
\alias{agg2pt}
\alias{agg2can}
\alias{agg2can_completeness}
\alias{dataset_format}
\title{Functions to further process COVID-19 data}
\usage{
convert_hr_names(d)

date_shift(d, n_days)

add_name_col(d, name)

drop_sub_regions(d, sr)

add_as_of_date(d, as_of_date)

report_pluck(d, name, val, out_col, geo)

report_recent(d)

max_date(d, max_date)

append_daily_d(d1, d2)

collate_datasets(val)

add_hr_col(d, name)

agg2pt(d, raw = FALSE)

agg2can(d)

agg2can_completeness(d)

dataset_format(d, geo = c("pt", "hr", "sub-hr"), digits)
}
\arguments{
\item{d}{The dataset to process.}

\item{n_days}{The number of days to shift dates forward by.}

\item{name}{The value name to add.}

\item{sr}{A vector of sub-regions to drop.}

\item{as_of_date}{The date to add as the "as-of date".}

\item{val}{The value to select.}

\item{out_col}{The name of the value column in the output dataset.}

\item{geo}{The geographic level of the data. One of "pt", "hr", "sub-hr".}

\item{max_date}{The maximum date to trim the output dataset to.}

\item{d1}{Dataset to append to. A cumulative value dataset.}

\item{d2}{Dataset being appended. A daily value dataset.}

\item{raw}{Is this function operating on raw data with only one of "value" or
"value_daily" columns?}

\item{digits}{The number of digits to round to for numeric values. If not
provided, no rounding will be done for numeric values.}
}
\description{
Functions to further process data downloaded via the `process_dataset`
function from `Covid19CanadaDataProcess`. Examples include normalizing
province and health region names and aggregating health region-level data up
to the province level.
}
