% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_funs.R
\name{process_funs}
\alias{process_funs}
\alias{process_prov_names}
\alias{process_hr_names}
\alias{process_agg2prov}
\alias{process_prov2hr}
\alias{process_cum_current}
\alias{process_collate}
\alias{sheets_load}
\alias{process_format_sheets}
\alias{process_format_dates}
\title{Functions to further process COVID-19 data}
\usage{
process_prov_names(d)

process_hr_names(d, prov, date_today = Sys.Date(), opt = NULL)

process_agg2prov(d)

process_prov2hr(d, prov)

process_cum_current(d)

process_collate(d, pattern)

sheets_load(id, sheet = NULL)

process_format_sheets(d, loc = c("prov", "hr"))

process_format_dates(...)
}
\arguments{
\item{d}{The dataset to process.}

\item{prov}{Two-letter province abbreviation (e.g., "ON").}

\item{date_today}{Today's date.}

\item{opt}{Optional. Other options passed to the function.}

\item{pattern}{The pattern of values to collate (e.g., "_cases_").}

\item{id}{The ID of the Google Sheet document to load.}

\item{sheet}{Optional. The specific sheet within the Google Sheet document to load.}

\item{loc}{Either "prov" or "hr".}

\item{...}{Other arguments.}
}
\description{
Functions to further process data downloaded via the `process_dataset`
function from `Covid19CanadaDataProcess`. Examples include normalizing
province and health region names and aggregating health region-level data up
to the province level.
}
